requirements.txt
google-generativeai>=0.8.0
PyPDF2>=3.0.0
PyYAML>=6.0
